Final Project "Data Science for engineers"
In this code I looked at 5 ambient variables in a power plant:
AT: Ambient Temperature
V: Exhaust Vacuum
AP: Ambient Pressure
RH: Relative Humidity

and our goal is to predict the best net energy output (EP) using appropriate statistical models.

I introduced a corrosion factor based on an equation with multiple assumptions to get an understanding of the wear and tear on the machines.

Our constraints were the OSHA regulations for temperature and humidity and trying to choose conditions for the lowest corrosion factors.