Data Analysis Project
This dataset is from a power plant which has monitored 5 variables.
https://archive.ics.uci.edu/ml/datasets/Combined+Cycle+Power+Plant

AT: Ambient Temperature
V: Exhaust Vacuum
AP: Ambient Pressure
RH: Relative Humidity
PE: Net Power Output

In this project I hope to model our data to predict the best ambient settings for the highest power output.

After finding the best power yield, we created a corrosion factor variable to give us an idea of the wear and tear on the machines in the plant. Note that this equation is highly flawed and is only used to see the trends.

The following link shows the equation: https://inis.iaea.org/collection/NCLCollectionStore/_Public/27/050/27050382.pdf

The constraints are keeping the ambient temperature and relative humidity up to OSHA standards and to try to minimize the corrosion factor.